requirements.txt
# Core dependencies
pyyaml>=6.0.1
pyvis>=0.3.2
networkx>=3.1
graphviz>=0.20.1
tenacity>=9.0.0
psutil>=5.9.0
cachetools>=5.3.0
astroid>=3.0.0
ollama>=0.1.0
rich>=13.7.0
imageio>=2.31.0  # For creating animated visualizations

# Error handling and logging
python-json-logger>=2.0.7
structlog>=23.1.0

# Development dependencies
black>=23.3.0
pylint>=2.17.0
mypy>=1.3.0
pytest>=7.3.1
pytest-cov>=4.1.0
memory-profiler>=0.61.0

# Optional dependencies for additional features
colorama>=0.4.6  # For colored terminal output
tqdm>=4.65.0    # For progress bars
